requirements.txt
protobuf==3.20.*
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
uvicorn~=0.17.6
fastapi~=0.76.0
etcd3~=0.12.0
PyQt5
kubernetes
pika==1.2.1
pyyaml
confluent-kafka
